Fix a few compilation warnings

+++ cuetools/cuetools-src/src/lib/time.c
@@ -19,7 +19,7 @@ void msf_frame_to_msf (long frame, int *m, int *s, int *f)
         frame /= 75;
         *s = frame % 60;	/* 0 <= seconds <= 59 */
         frame /= 60;
-        *m = frame;		/* 0 <= minutes */
+        *m = (int)frame;		/* 0 <= minutes */
 }
 
 void time_frame_to_msf (long frame, int *m, int *s, int *f)
@@ -28,7 +28,7 @@ void time_frame_to_msf (long frame, int *m, int *s, int *f)
 	frame /= 75;
 	*s = frame % 60;          /* 0 <= seconds <= 59 */
 	frame /= 60;
-	*m = frame;               /* 0 <= minutes */
+	*m = (int)frame;               /* 0 <= minutes */
 }
 
 /* print frame in mm:ss:ff format */
